engine/requirements.txt
# Core dependencies
manim>=0.18.0
requests>=2.31.0
python-dotenv>=1.0.0
pydantic>=2.5.0
anthropic>=0.18.0

# LLM integrations
google-generativeai>=0.3.0

# Development dependencies
pytest>=7.4.0
black>=23.12.0
mypy>=1.7.0
ruff>=0.1.0
